Remove executor.submit() around updateConfig due to generating concurrency exceptions
Executor.submit() was creating concurrency exceptions due to slowing write speed when updateConfig wrote to disk. As the config we want to write gets bigger, the more time we need to allow for writing. This may especially become evident later on when the tradeList gets sufficiently big or if the user's physical disk is too slow.