Add StatPanel sorting
This commit adds functionality to the sortBox already seen in the UI. It is implemented by running sortTradeList, which sorts the tradeList according to the specified user setting, just before the rebuild actually creates any of the panels present in the tradeList.

To gather information about when to sort for most recent, I've added latestActivityTime to the FlippingItem class. This is initiated with the time of creation and is later on set alongside any new trades that update the item. I've chosen to only allow complete trades to dictate the latestActivityTime to make sure the offers won't flicker between each other if the trades are ongoing.