Add initial separation of trade lists
This commit serves as the first step in facilitating the separation of trade lists between Flipping and StatsPanel. I've added two new booleans to OfferInfo (validFlippingOffer and validStatsOffer) that indicate whether the user has wanted these items deleted from the respective panel. If they've been deleted on the flipping panel, the statistical data of these items will still be present in the StatPanel along with the GE limit data, in case the user would happen to want this after it had been reset.

An offer is therefore not deleted after the user removes an item from the flipping panel (equally with the stat panel in a later commit). To make sure we don't create massive lists of objects, that could bring us into trouble with RL or the fact that the user might not be too thrilled that they have a massive config file present on their PC, I've implemented a way for the plugin to remove any offers that are deemed invalid for both the StatPanel and the FlippingPanel. Further optimization may be needed where tactics such as consolidation of standardized offers could prove sufficient to curb the majority of the concern behind memory usage.