Removed some SwingUtilities.invokeLater calls
These unnecessarily caused unresponsiveness in the UI. This will need
more testing since it likes to throw concurrency errors willy-nilly. If
this is the case, I recommend using invokeLater on the offending code
block, as it's most likely due to rebuild not being completely
finished before the code is executed involving the trade list.