Add Tracker Links (#62)
* Add tracker link support

Adds support and right-click options on items inside the flipping panel
to link tracker websites. Currently, this adds the official OSRS
Exchange website and Platinum Tokens.

There may be some issues with platinum tokens, as their slugging format
for item names is, at times, inconsistent. As of now, I've been able to
find an issue with Rune armour set (sk) and (lg), however there may be
more with issues. I don't believe this to be much of an issue, as the
default 404 page directs back to the homepage, meaning it won't be too
 disruptive for users to manually search the website.

* Refactor PopupMenus to UIUtils and add links to StatPanel

* Reduce platinum tokens URL failures

Added small exceptions to the parenthesis slugging to reduce a few cases
 of bad URLs on Platinum Tokens.

* Bump version