Refactor TradeActivityTimer and improve rebuilding
The timers now rebuild properly after each new offer has been submitted.
 It now uses ScriptPostFired events that subscribes to the GE redraw
 client script, which substantially improves timer rebuild reliability.

However, this meant the timers needed to allow for "duplicate" script
 events, as the redraw script can be called twice and sometimes thrice
 in a row without any clear indication as to why (Jagex spaghetti chefs
 are hard at work).

I've also implemented new thread logic as well, such that the timer
redraws are as smooth as possible. There's still some issues with it
that needs to be addressed but for now I think it's definitely fine.

Lastly, I've implemented a better way of tracking the time of the first
 trade such that the first item to be traded can be timed as well.