Add initial FlippingPanel redesign and toolbar
This commit marks the beginning of a flipping panel redesign and the
implementation of a toolbar. Currently, this just contains the resource
images, that I think look good enough to go on the user interface for
now.

Right now, the UI images and icons have a rather large resolution
(512x512 px) that I plan to cut down on once I finalize the look and
feel of the toolbar.

The toolbar itself also only has the visual components meaning no
functionality as of yet. I plan to also implement a FlippingItem
favorite system with the toolbar containing a button to select whether
their favorite items should appear at the top of the list, regardless of
the current sorting action on the list.